Generate Genetic Data

This function generates genetic data for a given number of samples. It simulates binary outcome variables, confounders, and genotypes of variants based on the specified parameters.

Usage

get_genetic_data(
  n,
  prob = 0.5,
  n_unif,
  min,
  max,
  prob_X = 0.5,
  n_norm,
  mean = 0,
  sd = 0.5
)

Arguments

n

Number of samples to generate data for.

prob

Probability for generating binary confounders.

n_unif

Number of uniform random variables to generate for minor allele frequency (MAF).

min

Minimum value for the uniform distribution used in generating MAF.

max

Maximum value for the uniform distribution used in generating MAF.

prob_X

Probability for the effect of confounders on the outcome.

n_norm

Number of normal random variables to generate for the genetic effect.

mean

Mean value for the normal distribution used in generating the genetic effect.

sd

Sd for the normal distribution used in generating the genetic effect.

Value

A data frame containing three variables:

Y: binary outcome variable. a vector of length n.

X: d confounders, either a nx1 vector or nxd matrix if d>1.

G: genotypes of varaints. a n*p matrix.

Examples

# To generate genetic data for 2000 samples:

set.seed(1234)
result <- get_genetic_data(n = 2000, n_unif=10, min=0.001,max=0.01, n_norm=10)
head(result)
#>   Y X G.1 G.2 G.3 G.4 G.5 G.6 G.7 G.8 G.9 G.10
#> 1 1 0   0   0   0   0   0   0   0   0   0    0
#> 2 0 1   0   0   0   0   0   0   0   0   0    0
#> 3 1 1   0   0   0   0   0   0   0   0   0    0
#> 4 1 1   0   0   0   0   0   0   0   0   0    0
#> 5 0 1   0   0   0   0   0   0   0   0   0    0
#> 6 0 1   0   0   0   0   0   0   0   0   0    0